pyproject.toml
[project]
name = "hmpps-github-discovery"
version = "1.1.1"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dockerfile-parse==2.0.1",
    "hmpps-sre-python-lib",
]

[tool.uv.sources]
hmpps-sre-python-lib = { git = "https://github.com/ministryofjustice/hmpps-sre-python-lib.git", rev = "feat/HEAT-1116_refactor_authentication" }
